Activité 6

+++ app/Http/Controllers/ProductController.php
@@ -4,49 +4,36 @@ namespace App\Http\Controllers;
 
 use App\Models\Product;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class ProductController extends Controller
 {
-    /**
-     * Display a listing of the resource.
-     */
     public function index()
     {
         $products = Product::with('user')->get();
-
         return view('products.index', compact('products'));
     }
 
-    /**
-     * Show the form for creating a new resource.
-     */
     public function create()
     {
         return view('products.create');
     }
 
-    /**
-     * Store a newly created resource in storage.
-     */
     public function store(Request $request)
     {
-        // Validation basique
         $validated = $request->validate([
             'name'  => ['required', 'string', 'max:255'],
             'price' => ['required', 'numeric', 'min:0'],
         ]);
 
-        // Création du produit pour l'utilisateur connecté
         Product::create([
             'name'     => $validated['name'],
             'price'    => $validated['price'],
             'user_id'  => $request->user()->id,
-            'is_public'=> $request->has('is_public'),
+            'is_public' => $request->has('is_public'),
         ]);
 
-        return redirect()
-            ->route('products.index')
-            ->with('status', 'Produit créé avec succès.');
+        return redirect()->route('products.index')->with('status', 'Produit créé avec succès.');
     }
 
     /**
@@ -54,6 +41,8 @@ class ProductController extends Controller
      */
     public function show(Product $product)
     {
+        Gate::authorize('view-product', $product); //
+
         return view('products.show', compact('product'));
     }
 
@@ -62,6 +51,8 @@ class ProductController extends Controller
      */
     public function edit(Product $product)
     {
+        Gate::authorize('manage-product', $product); //
+
         return view('products.edit', compact('product'));
     }
 
@@ -70,6 +61,8 @@ class ProductController extends Controller
      */
     public function update(Request $request,  Product $product)
     {
+        Gate::authorize('manage-product', $product); //
+
         $validated = $request->validate([
             'name'  => ['required', 'string', 'max:255'],
             'price' => ['required', 'numeric', 'min:0'],
@@ -91,6 +84,8 @@ class ProductController extends Controller
      */
     public function destroy(Product $product)
     {
+        Gate::authorize('manage-product', $product); //
+
         $product->delete();
 
         return redirect()
